+++ src/lib/components/modal/select-contract/SelectContract.tsx
@@ -21,10 +21,10 @@ import { MdChevronLeft, MdList, MdSwapHoriz } from "react-icons/md";
 
 import { useCelatoneApp } from "lib/app-provider";
 import { DEFAULT_RPC_ERROR } from "lib/data";
-import { useContractStore, useEndpoint } from "lib/hooks";
+import { useContractStore, useEndpoint, useValidateAddress } from "lib/hooks";
 import { useInstantiatedByMe } from "lib/model/contract";
 import { queryContract } from "lib/services/contract";
-import type { ContractAddr, RpcContractError } from "lib/types";
+import type { ContractAddr, RpcQueryError } from "lib/types";
 
 import { AllContractLists } from "./AllContractLists";
 import { ContractListDetail } from "./ContractListDetail";
@@ -43,6 +43,7 @@ export const SelectContract = ({
   } = useCelatoneApp();
   const { isOpen, onOpen, onClose } = useDisclosure();
   const [listSlug, setListSlug] = useState("");
+  const { validateContractAddress } = useValidateAddress();
 
   const [searchContract, setSearchContract] = useState<ContractAddr>(
     "" as ContractAddr
@@ -79,8 +80,8 @@ export const SelectContract = ({
       onSuccess() {
         onSelectThenClose(searchContract);
       },
-      onError(err: AxiosError<RpcContractError>) {
-        setInvalid(err.response?.data.error || DEFAULT_RPC_ERROR);
+      onError(err: AxiosError<RpcQueryError>) {
+        setInvalid(err.response?.data.message || DEFAULT_RPC_ERROR);
       },
     }
   );
@@ -131,7 +132,9 @@ export const SelectContract = ({
                   isDisabled={searchContract.length === 0}
                   isLoading={isFetching || isRefetching}
                   onClick={() => {
-                    refetch();
+                    const err = validateContractAddress(searchContract);
+                    if (err !== null) setInvalid(err);
+                    else refetch();
                   }}
                 >
                   Submit